test(datetime): migrate set-value
